requirements.txt
accelerate==0.33.0
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
contourpy==1.3.0
controlnet-aux==0.0.9
cycler==0.12.1
diffusers==0.30.0
einops==0.8.0
executing==2.0.1
filelock==3.15.4
fonttools==4.54.0
fsspec==2024.6.1
ftfy==6.2.3
hf-transfer==0.1.8
huggingface-hub==0.24.5
idna==3.7
imageio==2.35.1
importlib-metadata==8.2.0
importlib-resources==6.4.5
jinja2==3.1.4
joblib==1.4.2
jupyter-client==7.4.9
kiwisolver==1.4.7
lazy-loader==0.4
lpips==0.1.4
markupsafe==2.1.5
matplotlib==3.7.1
mpmath==1.3.0
mypy-extensions==1.0.0
networkx==3.2.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.20
nvidia-nvtx-cu12==12.1.105
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
packaging==24.1
pillow==10.4.0
platformdirs==4.2.2
pygments==2.18.0
pyparsing==3.1.4
pyrallis==0.3.1
regex==2024.7.24
requests==2.32.3
safetensors==0.4.4
scikit-image==0.24.0
scikit-learn==1.1.1
scipy==1.13.1
six==1.16.0
sympy==1.13.2
threadpoolctl==3.5.0
tifffile==2024.8.28
timm==0.6.7
tokenizers==0.19.1
torch==2.3.0
torchaudio==2.3.0
torchvision==0.18.0
transformers==4.44.0
triton==2.3.0
types-python-dateutil==2.9.0.20240821
typing-extensions==4.12.2
typing-inspect==0.9.0
urllib3==1.26.14
virtualenv==20.26.3
watchdog==5.0.0
xformers==0.0.26.post1
zipp==3.20.0